requirements.txt
transformers==4.35.2
flask==3.0.0
pydicom==2.4.2
opencv-python==4.8.1.78
Pillow==10.1.0
gunicorn==20.1.0
flask-swagger-ui==4.11.1
boto3==1.26.119
